+++ src/account/action/moneyTransfer.ts
@@ -2,23 +2,19 @@ import { AccountType } from '../../accountType';
 import * as ActionFactory from '../../action';
 import { ActionType } from '../../actionType';
 import { IMonetaryAmount } from '../../monetaryAmount';
-import { IThing } from '../../thing';
 import { AccountTransactionType } from '../transactionType';
 
-export type IAgent = ActionFactory.IParticipant;
-export type IRecipient = ActionFactory.IParticipant;
+export type IAgent = Pick<ActionFactory.IParticipant, 'name' | 'typeOf'>;
+export type IRecipient = Pick<ActionFactory.IParticipant, 'name' | 'typeOf'>;
 /**
  * 口座以外の匿名ロケーション
  */
-export interface IAnonymousLocation extends Pick<IThing, 'identifier' | 'name'> {
+export interface IAnonymousLocation {
+    name?: string;
     /**
      * ロケーションタイプ
      */
     typeOf: string;
-    /**
-     * ロケーションID
-     */
-    id?: string;
 }
 /**
  * 口座
@@ -27,8 +23,9 @@ export interface IAccount {
     typeOf: AccountType.Account;
     /**
      * 通貨
+     * 不要なので廃止(2023-02-16~)
      */
-    accountType: string;
+    // accountType: string;
     /**
      * 口座番号
      */
@@ -42,9 +39,6 @@ export interface IAccount {
  * 転送元あるいは転送先
  */
 export type ILocation = IAnonymousLocation | IAccount;
-export type IObject = any;
-export type IResult = any;
-export type IPotentialActions = any;
 /**
  * アクションの目的
  * ここでは、取引が目的となる
@@ -65,11 +59,18 @@ export interface IPurpose {
     /**
      * 取引識別子
      */
-    identifier?: string;
+    // identifier?: string;
 }
-export type IAmount = Pick<IMonetaryAmount, 'typeOf' | 'currency' | 'value'>;
-export interface IAttributes extends ActionFactory.IAttributes<ActionType.MoneyTransfer, IObject, IResult> {
-    identifier?: string;
+// 最適化(2023-02-16~)
+// export type IAmount = Pick<IMonetaryAmount, 'typeOf' | 'currency' | 'value'>;
+export type IAmount = Pick<IMonetaryAmount, 'typeOf' | 'value'>;
+export interface IAttributes extends Pick<
+    ActionFactory.IAttributes<ActionType.MoneyTransfer, never, never>,
+    'project' | 'description'
+> {
+    agent: IAgent;
+    recipient?: IRecipient;
+    // identifier?: string;
     typeOf: ActionType.MoneyTransfer;
     /**
      * どんな取引によって発生した転送アクションか
@@ -88,4 +89,13 @@ export interface IAttributes extends ActionFactory.IAttributes<ActionType.MoneyT
      */
     toLocation: ILocation;
 }
-export type IAction = ActionFactory.IAction<IAttributes>;
+export type IAction = Pick<
+    ActionFactory.IAction<Omit<IAttributes, 'agent' | 'recipient'> & {
+        agent: ActionFactory.IParticipant;
+        object: never;
+    }>,
+    'actionStatus' | 'amount' | 'description' | 'endDate' | 'fromLocation' | 'project' | 'purpose' | 'startDate' | 'toLocation' | 'typeOf' | 'id'
+> & {
+    agent: IAgent;
+    recipient?: IRecipient;
+};