+++ src/event/screeningEvent.ts
@@ -172,12 +172,13 @@ export type ISuperEvent = Omit<ScreeningEventSeriesFactory.IEvent, 'eventStatus'
 export import IName = AnyEventFactory.IName;
 /**
  * イベント属性
+ * workPerformed->superEvent.workPerformed完全移行(2023-02-20~)
  */
-export interface IAttributes extends Omit<AnyEventFactory.IAttributes, 'offers' | 'typeOf'> {
+export interface IAttributes extends Omit<AnyEventFactory.IAttributes, 'offers' | 'typeOf' | 'workPerformed'> {
     /**
      * コンテンツ
      */
-    workPerformed?: IWorkPerformed;
+    // workPerformed?: IWorkPerformed;
     /**
      * 親イベント
      * 施設コンテンツに相当